# docs/swagger.yaml
basePath: /
definitions:
...
  utils.BadRequest:
    properties:
      error:
        example: string
        type: string
      message:
        example: message
        type: string
      result:
        example: "null"
        type: string
      status:
        example: 400
        type: integer
      success:
        example: false
        type: boolean
    type: object
...
  utils.InternalServerError:
    properties:
      error:
        example: message
        type: string
      message:
        example: message
        type: string
      result:
        example: "null"
        type: string
      status:
        example: 500
        type: integer
      success:
        example: false
        type: boolean
    type: object
...
  utils.Success:
    properties:
      error:
        example: "null"
        type: string
      message:
        example: message
        type: string
      result:
        example: message
        type: string
      status:
        example: 200
        type: integer
      success:
        example: true
        type: boolean
    type: object
  utils.Welcome:
    properties:
      otp:
        example: 1234
        type: integer
      to:
        example: mail@example.com
        type: string
    type: object
info:
  contact: {}
  description: Documentation for Postmail
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Postmail API
  version: "1.0"
paths:
...
  /send/template/welcome:
    post:
      consumes:
      - application/json
      description: Using template Welcome
      parameters:
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/utils.Welcome'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/utils.Success'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.BadRequest'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.InternalServerError'
      security:
      - X-Auth-Key: []
      summary: Send email template - Welcome
      tags:
      - Send Template
securityDefinitions:
  X-Auth-Key:
    description: Input your token authorized
    in: header
    name: X-Auth-Key
    type: apiKey
swagger: "2.0"